Move third party consent components to sub directory
Make it clearer which files belong to the feature and which are
considered public interface. Do not export `ThirdPartyConsentProvider`
in `pageflow-scrolled/frontend` since it is only supposed to be used
in root providers - not content elements.

REDMINE-17863
