+++ entry_types/scrolled/package/src/frontend/thirdPartyConsent/OptOut.module.css
@@ -22,7 +22,7 @@
 }
 
 .icon {
-  composes: unstyledButton from './utils.module.css';
+  composes: unstyledButton from '../utils.module.css';
   position: relative;
   background-color: rgb(0, 55, 90);
 }
