+++ entry_types/scrolled/package/src/frontend/thirdPartyConsent/index.js
@@ -1,0 +1,2 @@
+export {ThirdPartyConsentProvider} from './ThirdPartyConsentProvider';
+export {ThirdPartyConsent} from './ThirdPartyConsent';
